refactor: ダッシュボードのレイアウトを Dashboard2 デザインに統合
- HeroSummary (次にやること提示) / RecruitmentBoard (1行集約カード) / StaffRoster (アバター付き行) を Dashboard/ に統合
- StaffRoster の API を onEdit(staff) / onDelete(staff) に変更し編集・削除メニューを内蔵
- 旧レイアウト (ShopInfoBar / RecruitmentSection / RecruitmentCard / StaffSection / StaffListItem / RecruitmentStatusBadge / LoadMoreButton) を削除
- Dashboard2 ディレクトリを削除
- 初回ログイン時のヒーロー (WelcomeHero) を Storybook 用に保持
- E2E DashboardPage をラベル変更に合わせて更新 (シフトを見る / 新しく募集する / もっと見る)
