+++ src/components/features/Dashboard/RecruitmentBoard/index.stories.tsx
@@ -1,25 +1,25 @@
 import { Stack, Text } from "@chakra-ui/react";
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { mockStaffs, mockStaffsMany } from "@/src/components/features/Dashboard/storyMocks";
-import { StaffRoster } from ".";
+import { mockRecruitments } from "@/src/components/features/Dashboard/storyMocks";
+import { RecruitmentBoard } from ".";
 
 const meta = {
-  title: "Features/Dashboard2/StaffRoster",
-  component: StaffRoster,
+  title: "Features/Dashboard/RecruitmentBoard",
+  component: RecruitmentBoard,
   parameters: {
     layout: "padded",
   },
-} satisfies Meta<typeof StaffRoster>;
+} satisfies Meta<typeof RecruitmentBoard>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Variants: Story = {
   args: {
-    staffs: mockStaffs,
+    recruitments: mockRecruitments,
     status: "Exhausted",
-    onAddClick: () => {},
-    onMenuClick: () => {},
+    onCreateClick: () => {},
+    onOpenShiftBoard: () => {},
     onLoadMore: () => {},
   },
   render: () => (
@@ -28,11 +28,11 @@ export const Variants: Story = {
         <Text fontSize="xs" fontWeight="semibold" color="fg.muted" letterSpacing="0.08em" textTransform="uppercase">
           データあり
         </Text>
-        <StaffRoster
-          staffs={mockStaffs}
+        <RecruitmentBoard
+          recruitments={mockRecruitments}
           status="Exhausted"
-          onAddClick={() => {}}
-          onMenuClick={() => {}}
+          onCreateClick={() => {}}
+          onOpenShiftBoard={() => {}}
           onLoadMore={() => {}}
         />
       </Stack>
@@ -40,11 +40,11 @@ export const Variants: Story = {
         <Text fontSize="xs" fontWeight="semibold" color="fg.muted" letterSpacing="0.08em" textTransform="uppercase">
           もっと見る ありの状態
         </Text>
-        <StaffRoster
-          staffs={mockStaffsMany}
+        <RecruitmentBoard
+          recruitments={mockRecruitments.slice(0, 2)}
           status="CanLoadMore"
-          onAddClick={() => {}}
-          onMenuClick={() => {}}
+          onCreateClick={() => {}}
+          onOpenShiftBoard={() => {}}
           onLoadMore={() => {}}
         />
       </Stack>
@@ -52,11 +52,11 @@ export const Variants: Story = {
         <Text fontSize="xs" fontWeight="semibold" color="fg.muted" letterSpacing="0.08em" textTransform="uppercase">
           空の状態
         </Text>
-        <StaffRoster
-          staffs={[]}
+        <RecruitmentBoard
+          recruitments={[]}
           status="Exhausted"
-          onAddClick={() => {}}
-          onMenuClick={() => {}}
+          onCreateClick={() => {}}
+          onOpenShiftBoard={() => {}}
           onLoadMore={() => {}}
         />
       </Stack>
